+++ apps/api/src/index.ts
@@ -1,7 +1,7 @@
 import express, { Express } from 'express'
 import cors from 'cors'
 import { cronjob } from './database/Splits'
-import { userIdMiddleware } from './middleware/middleware'
+import { userIdMiddleware } from './v3/middleware/middleware'
 
 const app: Express = express()
 const PORT = 8080
@@ -43,7 +43,19 @@ app.use(`/${VERSION}/weights`, weightsRouter)
 import workoutsRouter from './routes/Workouts'
 app.use(`/${VERSION}/workouts`, workoutsRouter)
 
-import headerRouter from './routes/Header'
-app.use(`/${VERSION}/header`, headerRouter)
+import exercises from './v3/routes/Exercises'
+app.use(`/v3/exercises`, exercises)
+
+import splits from './v3/routes/Splits'
+app.use(`/v3/splits`, splits)
+
+import weights from './v3/routes/Weights'
+app.use(`/v3/weights`, weights)
+
+import workouts from './v3/routes/Workouts'
+app.use(`/v3/workouts`, workouts)
+
+import header from './v3/routes/Header'
+app.use(`/v3/header`, header)
 
 export default app
