+++ apps/api/src/routes/Workouts.ts
@@ -1,29 +1,28 @@
-import express, { Request, Response, Router } from 'express'
-const router: Router = express.Router()
+import express, { Request, Response } from 'express'
+const router = express.Router()
 
-import { getCurrentWorkoutCount, getWorkout, insertWorkout } from '../database/Workouts'
+import { getCurrentWorkoutNumber, getWorkout, insertWorkout } from '../database/Workouts'
 
 router.get('/count', async (req: Request, res: Response) => {
   const userId = req.query.userId as string
-
-  const { statusCode, ...response } = await getCurrentWorkoutCount(userId)
-  res.status(statusCode).json({ ...response })
+  const rows = await getCurrentWorkoutNumber(userId)
+  res.json(rows[0])
 })
 
 router.post('/', async (req: Request, res: Response) => {
   const userId = req.query.userId as string
   const workout = req.body.workout
-
-  const { statusCode, ...response } = await insertWorkout(userId, workout)
-  res.status(statusCode).json({ ...response })
+  await insertWorkout(userId, workout)
+  res.sendStatus(201)
 })
 
 router.get('/:workoutId', async (req: Request, res: Response) => {
   const userId = req.query.userId as string
   const workoutId = req.params.workoutId
+  const rows = await getWorkout(userId, workoutId)
 
-  const { statusCode, ...response } = await getWorkout(userId, workoutId)
-  res.status(statusCode).json({ ...response })
+  if (rows.length === 0) res.status(400).json({ error: 'No matching records' })
+  else res.status(200).json(rows)
 })
 
 export default router
