+++ apps/api/src/v3/routes/Weights.ts
@@ -1,5 +1,5 @@
-import express, { Request, Response } from 'express'
-const router = express.Router()
+import express, { Request, Response, Router } from 'express'
+const router: Router = express.Router()
 
 import {
   createEntry,
@@ -11,37 +11,39 @@ import {
 
 router.get('/', async (req: Request, res: Response) => {
   const userId = req.query.userId as string
-  const rows = await getAllWeight(userId)
-  res.status(200).json(rows)
+
+  const { statusCode, ...response } = await getAllWeight(userId)
+  res.status(statusCode).json({ ...response })
 })
 
 router.post('/', async (req: Request, res: Response) => {
   const userId = req.query.userId as string
   const { weight, date } = req.body
-  await createEntry(userId, weight, date)
-  res.sendStatus(201)
+
+  const { statusCode, ...response } = await createEntry(userId, weight, date)
+  res.status(statusCode).json({ ...response })
 })
 
 router.get('/current', async (req: Request, res: Response) => {
   const userId = req.query.userId as string
-  const row = await getCurrentWeight(userId)
-  res.status(200).json(row[0])
+  const { statusCode, ...response } = await getCurrentWeight(userId)
+  res.status(statusCode).json({ ...response })
 })
 
 router.get('/:id', async (req: Request, res: Response) => {
   const userId = req.query.userId as string
   const id = req.params.id
-  const row = await getWeight(userId, id)
 
-  if (row.length == 0) res.status(400).json({ error: 'No matching records' })
-  else res.status(200).json(row)
+  const { statusCode, ...response } = await getWeight(userId, id)
+  return res.status(statusCode).json({ ...response })
 })
 
 router.delete('/:id', async (req: Request, res: Response) => {
   const userId = req.query.userId as string
   const id = req.params.id
-  await deleteEntry(userId, id)
-  res.sendStatus(204)
+
+  const { statusCode, ...response } = await deleteEntry(userId, id)
+  return res.status(statusCode).json({ ...response })
 })
 
 export default router
